requirements.txt
flask==2.0.1
flask_login==0.5.0
flask_migrate==3.1.0
flask_wtf==0.15.1
WTForms==2.3.3
flask_sqlalchemy==2.5.1
sqlalchemy==1.4.23
email_validator==1.1.3
python-decouple==3.4
gunicorn==20.1.0
jinja2~=3.0.3
flask-restx==0.5.1
pandas==1.3.4
nltk==3.6.5
wordcloud==1.6.0
numpy==1.21.0
matplotlib==3.1.2
Werkzeug~=2.0.0
